Feature/mcp 1962 disable endpoints (#524)
* separate end points to be disabled

* separate controllers to be disabled

* separate controller tests

* code cleaning

* code cleaning

* remove dev controller from qa

* remove unused
