+++ api/src/main/java/gov/va/vro/api/resources/VroResource.java
@@ -6,10 +6,8 @@ import gov.va.vro.api.model.MetricsProcessingException;
 import gov.va.vro.api.requests.GeneratePdfRequest;
 import gov.va.vro.api.requests.HealthDataAssessmentRequest;
 import gov.va.vro.api.responses.ClaimMetricsResponse;
-import gov.va.vro.api.responses.FetchClaimsResponse;
 import gov.va.vro.api.responses.FullHealthDataAssessmentResponse;
 import gov.va.vro.api.responses.GeneratePdfResponse;
-import gov.va.vro.api.responses.HealthDataAssessmentResponse;
 import io.micrometer.core.annotation.Timed;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -40,48 +38,11 @@ import javax.validation.Valid;
 @Timed
 public interface VroResource {
   @Operation(
-      summary = "Health data assessment",
-      description = "Provides health data assessment for the claim")
-  @PostMapping("/health-data-assessment")
-  @ResponseStatus(HttpStatus.CREATED)
-  @ApiResponses(
-      value = {
-        @ApiResponse(responseCode = "201", description = "Successful Request"),
-        @ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized",
-            content = @Content(schema = @Schema(hidden = true))),
-        @ApiResponse(
-            responseCode = "500",
-            description = "Data Access Server Error",
-            content = @Content(schema = @Schema(hidden = true))),
-        @ApiResponse(
-            responseCode = "404",
-            description = "No evidence found",
-            content =
-                @Content(
-                    mediaType = "application/json",
-                    examples =
-                        @ExampleObject(
-                            value = "{claimSubmissionId: 1234, message = No evidence found}")))
-      })
-  @Timed(value = "health-data-assessment")
-  @Tag(name = "Health Assessment")
-  ResponseEntity<HealthDataAssessmentResponse> postHealthAssessment(
-      @Parameter(
-              description = "Claim for which health data assessment requested",
-              required = true,
-              schema = @Schema(implementation = HealthDataAssessmentRequest.class))
-          @Valid
-          @RequestBody
-          HealthDataAssessmentRequest claim)
-      throws RequestValidationException, ClaimProcessingException;
-
-  @Operation(
-      summary = "Evidence pdf generation launch",
+      summary = "Generates Evidence PDF",
       description =
-          "Posts health data for a specific patient and diagnostic code to "
-              + "launch evidence pdf generation")
+          "This endpoint generates the Evidence PDF for a specific patient and a diagnostic "
+              + "code. The evidence pdf will be available from 'GET evidence-pdf' endpoint "
+              + "using claim submission id.")
   @PostMapping("/evidence-pdf")
   @ApiResponses(
       value = {
@@ -108,10 +69,11 @@ public interface VroResource {
       throws RequestValidationException, ClaimProcessingException;
 
   @Operation(
-      summary = "Generated evidence pdf download",
+      summary = "Downloads the generated Evidence PDF",
       description =
-          "Downloads the evidence pdf previously generated by the posted"
-              + " patient health data and diagnostic code")
+          "This endpoint downloads the Evidence PDF previously generated by the "
+              + "'POST evidence-pdf' endpoint. "
+              + "The Evidence PDF is identified by the claim submission id.")
   @GetMapping("/evidence-pdf/{claimSubmissionId}")
   @ApiResponses(
       value = {
@@ -135,8 +97,11 @@ public interface VroResource {
       throws RequestValidationException, ClaimProcessingException;
 
   @Operation(
-      summary = "Full health data assessment",
-      description = "Provides full health data assessment for a claim")
+      summary = "Provides health data assessment",
+      description =
+          "This endpoint provides health data assessment for a Veteran claim "
+              + "in the form of patient medical data relevant to the specific diagnostic code. "
+              + "Claim id is only used for tracking purposes.")
   @PostMapping("/full-health-data-assessment")
   @ResponseStatus(HttpStatus.CREATED)
   @ApiResponses(
@@ -173,32 +138,10 @@ public interface VroResource {
       throws RequestValidationException, ClaimProcessingException;
 
   @Operation(
-      summary = "Gets all claims stored in vro database.",
+      summary = "Retrieves metrics on the previously processed claims",
       description =
-          "Retrieves all claims from vro db and displays claimSubmissionId,"
-              + " veteran ICN, and contention diagnostic codes.")
-  @ApiResponses(
-      value = {
-        @ApiResponse(responseCode = "201", description = "Successful"),
-        @ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized",
-            content = @Content(schema = @Schema(hidden = true))),
-        @ApiResponse(
-            responseCode = "500",
-            description = "Fetch Claims Server Error",
-            content = @Content(schema = @Schema(hidden = true)))
-      })
-  @GetMapping("/fetch-claims")
-  @ResponseStatus(HttpStatus.OK)
-  @Timed(value = "fetch-claims")
-  @Tag(name = "Claim Metrics")
-  ResponseEntity<FetchClaimsResponse> fetchClaims()
-      throws RequestValidationException, ClaimProcessingException;
-
-  @Operation(
-      summary = "Retrieves metrics on claims stored in abd-vro database",
-      description = "Returns number of claims in abd-vro databse")
+          "This endpoint provides metrics on the previously processed claims. "
+              + "Currently only the number of the processed claims is provided.")
   @ApiResponses(
       value = {
         @ApiResponse(responseCode = "201", description = "Successful"),
