+++ app/src/test/java/gov/va/vro/VroControllerTest.java
@@ -12,15 +12,15 @@ import gov.va.vro.api.responses.FetchPdfResponse;
 import gov.va.vro.api.responses.FullHealthDataAssessmentResponse;
 import gov.va.vro.api.responses.GeneratePdfResponse;
 import gov.va.vro.api.responses.HealthDataAssessmentResponse;
+import gov.va.vro.config.AppTestConfig;
+import gov.va.vro.config.AppTestUtil;
 import gov.va.vro.controller.exception.ClaimProcessingError;
 import gov.va.vro.model.AbdEvidence;
 import gov.va.vro.model.VeteranInfo;
 import gov.va.vro.persistence.model.ClaimEntity;
-import gov.va.vro.persistence.repository.ClaimRepository;
 import gov.va.vro.service.provider.camel.FunctionProcessor;
 import gov.va.vro.service.provider.camel.PrimaryRoutes;
 import gov.va.vro.service.spi.model.Claim;
-import lombok.SneakyThrows;
 import org.apache.camel.CamelContext;
 import org.apache.camel.EndpointInject;
 import org.apache.camel.builder.Builder;
@@ -30,6 +30,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.annotation.DirtiesContext;
@@ -42,17 +43,13 @@ import java.util.function.Function;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
+@Import(AppTestConfig.class)
 @CamelSpringBootTest
 class VroControllerTest extends BaseControllerTest {
 
-  private final ObjectMapper objectMapper = new ObjectMapper();
-
-  @Autowired private ClaimRepository claimRepository;
+  @Autowired private AppTestUtil util;
 
   @Autowired protected CamelContext camelContext;
-
-  @EndpointInject("mock:claim-submit")
-  private MockEndpoint mockSubmitClaimEndpoint;
 
   @EndpointInject("mock:claim-submit-full")
   private MockEndpoint mockFullHealthEndpoint;
@@ -65,87 +62,6 @@ class VroControllerTest extends BaseControllerTest {
 
   @Value("classpath:test-data/pdf-generator-input-01.json")
   private Resource pdfGeneratorInput01;
-
-  @Test
-  @DirtiesContext
-  void postHealthAssessment() throws Exception {
-
-    // intercept the original endpoint, skip it and replace it with the mock
-    // endpoint
-    adviceWith(
-        camelContext,
-        "claim-submit",
-        route ->
-            route
-                .interceptSendToEndpoint(
-                    "rabbitmq:claim-submit-exchange"
-                        + "?queue=claim-submit"
-                        + "&routingKey=code.1701&requestTimeout=60000")
-                .skipSendToOriginalEndpoint()
-                .to("mock:claim-submit"));
-    // The mock endpoint returns a valid response
-    mockSubmitClaimEndpoint.whenAnyExchangeReceived(
-        FunctionProcessor.<Claim, String>fromFunction(claim -> claimToResponse(claim, true)));
-
-    HealthDataAssessmentRequest request = new HealthDataAssessmentRequest();
-    request.setClaimSubmissionId("1234");
-    request.setVeteranIcn("icn");
-    request.setDiagnosticCode("1701");
-
-    var responseEntity1 =
-        post("/v1/health-data-assessment", request, HealthDataAssessmentResponse.class);
-
-    assertEquals(HttpStatus.CREATED, responseEntity1.getStatusCode());
-    HealthDataAssessmentResponse response1 = responseEntity1.getBody();
-    assertNotNull(response1);
-    assertEquals(request.getDiagnosticCode(), response1.getDiagnosticCode());
-    assertEquals(request.getVeteranIcn(), response1.getVeteranIcn());
-
-    // Now submit an existing claim:
-    var responseEntity2 =
-        post("/v1/health-data-assessment", request, HealthDataAssessmentResponse.class);
-    assertEquals(HttpStatus.CREATED, responseEntity2.getStatusCode());
-    HealthDataAssessmentResponse response2 = responseEntity2.getBody();
-    assertNotNull(response2);
-    assertEquals(request.getDiagnosticCode(), response2.getDiagnosticCode());
-    assertEquals(request.getVeteranIcn(), response2.getVeteranIcn());
-
-    // side effect: claim is created in the DB
-    Optional<ClaimEntity> claimEntityOptional =
-        claimRepository.findByClaimSubmissionIdAndIdType("1234", "va.gov-Form526Submission");
-    assertTrue(claimEntityOptional.isPresent());
-  }
-
-  @Test
-  @DirtiesContext
-  void claimSubmit_missing_evidence() throws Exception {
-    adviceWith(
-        camelContext,
-        "claim-submit",
-        route ->
-            route
-                .interceptSendToEndpoint(
-                    "rabbitmq:claim-submit-exchange"
-                        + "?queue=claim-submit"
-                        + "&routingKey=code.1701&requestTimeout=60000")
-                .skipSendToOriginalEndpoint()
-                .to("mock:claim-submit"));
-
-    mockSubmitClaimEndpoint.whenAnyExchangeReceived(
-        FunctionProcessor.<Claim, String>fromFunction(claim -> claimToResponse(claim, false)));
-
-    HealthDataAssessmentRequest request = new HealthDataAssessmentRequest();
-    request.setClaimSubmissionId("1234");
-    request.setVeteranIcn("icn");
-    request.setDiagnosticCode("1701");
-
-    var responseEntity = post("/v1/health-data-assessment", request, ClaimProcessingError.class);
-    assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode());
-    var claimProcessingError = responseEntity.getBody();
-    assertNotNull(claimProcessingError);
-    assertEquals("No evidence found.", claimProcessingError.getMessage());
-    assertEquals("1234", claimProcessingError.getClaimSubmissionId());
-  }
 
   @Test
   @DirtiesContext
@@ -177,7 +93,7 @@ class VroControllerTest extends BaseControllerTest {
                 .to("mock:claim-submit-full"));
     // The mock endpoint returns a valid response
     mockFullHealthEndpoint.whenAnyExchangeReceived(
-        FunctionProcessor.<Claim, String>fromFunction(claim -> claimToResponse(claim, true)));
+        FunctionProcessor.<Claim, String>fromFunction(claim -> util.claimToResponse(claim, true)));
 
     HealthDataAssessmentRequest request = new HealthDataAssessmentRequest();
     request.setClaimSubmissionId("1234");
@@ -234,7 +150,7 @@ class VroControllerTest extends BaseControllerTest {
                 .to("mock:claim-submit-full"));
 
     mockFullHealthEndpoint.whenAnyExchangeReceived(
-        FunctionProcessor.<Claim, String>fromFunction(claim -> claimToResponse(claim, false)));
+        FunctionProcessor.<Claim, String>fromFunction(claim -> util.claimToResponse(claim, false)));
 
     HealthDataAssessmentRequest request = new HealthDataAssessmentRequest();
     request.setClaimSubmissionId("1234");
@@ -301,26 +217,9 @@ class VroControllerTest extends BaseControllerTest {
 
     mockFetchPdfEndpoint.whenAnyExchangeReceived(
         FunctionProcessor.fromFunction(
-            (Function<Object, Object>) o -> toJsonString(fetchPdfResponse)));
+            (Function<Object, Object>) o -> util.toJsonString(fetchPdfResponse)));
 
     var response = get("/v1/evidence-pdf/1234", null, String.class);
     assertNotNull(response);
-  }
-
-  @SneakyThrows
-  private String toJsonString(Object o) {
-    return objectMapper.writeValueAsString(o);
-  }
-
-  @SneakyThrows
-  private String claimToResponse(Claim claim, boolean evidence) {
-    var response = new HealthDataAssessmentResponse();
-    response.setDiagnosticCode(claim.getDiagnosticCode());
-    response.setVeteranIcn(claim.getVeteranIcn());
-    response.setErrorMessage("I am not a real endpoint.");
-    if (evidence) {
-      response.setEvidence(new AbdEvidence());
-    }
-    return objectMapper.writeValueAsString(response);
   }
 }
