+++ domain/src/main/java/gov/va/vro/model/AbdEvidence.java
@@ -3,6 +3,7 @@ package gov.va.vro.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -13,10 +14,16 @@ import java.util.List;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AbdEvidence {
+  @Schema(description = "List of relevant medications")
   private List<AbdMedication> medications;
+
+  @Schema(description = "List of relevant conditions")
   private List<AbdCondition> conditions;
+
+  @Schema(description = "List of relevant procedures")
   private List<AbdProcedure> procedures;
 
+  @Schema(description = "List of relevant blood pressures")
   @JsonProperty("bp_readings")
   private List<AbdBloodPressure> bloodPressures;
 }
